fix: #6 filter path is not configured at startup of codestart
The WicketFilter requires to know what the filter path is, but that
information is not available in a portable way through the servlet spec
so it is required to be set through configuration or programmatically.

This issue is resolved to always use `/*` as the filter path, as this is
typically what you want for a Wicket web-application.

At a later time we will try to make this more flexible to be able to
provide a REST api and a Wicket endpoint from the same application, e.g.

- /api/* for REST
- /app/* for Wicket

where `/` redirects to `/app` or so.

For now, for release 1.0 this is sufficient.

Fixes #6